requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
dnspython>=2.1.0
Flask>=2.0.0
cryptography>=3.4.0 